backend/requirements.txt
asgiref==3.3.1
Django==3.1.5
djangorestframework==3.12.2
Pillow==8.1.0
psycopg2-binary==2.8.6
pytz==2020.5
sqlparse==0.4.1
